renaming and pathlibbing
